requirements.txt
altgraph==0.17.4
astroid==3.2.3
av==12.0.0
black==24.4.2
blinker==1.8.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
coloredlogs==15.0.1
ctranslate2==4.3.1
dill==0.3.8
faster-whisper==1.0.2
filelock==3.14.0
Flask==3.0.3
flatbuffers==24.3.25
fsspec==2024.5.0
huggingface-hub==0.23.0
humanfriendly==10.0
idna==3.7
isort==5.13.2
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
mccabe==0.7.0
mpmath==1.3.0
mypy-extensions==1.0.0
numpy==1.26.4
onnxruntime==1.18.0
packaging==24.0
pathspec==0.12.1
platformdirs==4.2.2
protobuf==5.26.1
PyAudio==0.2.14
pyinstaller==6.6.0
pyinstaller-hooks-contrib==2024.6
pylint==3.2.5
python-dotenv==1.0.1
PyYAML==6.0.1
requests==2.31.0
sympy==1.12
tokenizers==0.19.1
tomli==2.0.1
tomlkit==0.13.0
tqdm==4.66.4
typing_extensions==4.11.0
urllib3==2.2.1
waitress==3.0.0
Werkzeug==3.0.3
